flights/requirements.txt
Flask==3.0.0
databricks-sql-connector==3.0.2
databricks-sdk==0.17.0
pandas==2.1.4
pyarrow==14.0.1

